Worked with shaders and movement animations incl. pivot
The animations are still a bit weird but I think they are at least
passable to not annoy me during development too much. The Xray still
needs a tiny bit of tweaking as well. However, I am moving on to
working with something else for now...maybe bringing back the day-
night cycle?