Fixed various movement issues, still has remaining probblems
The main remaining problem seems to be around the smoothing of the
animation speeds so they could blend together. I also had to do
a lerp hack on 135deg rotations since otherwise they do not land
in the directionInt 0 zone which is required for directionInt to
actually update in animator...an ugly loop of 135deg resulted.